refactor: post word and input separated
